refactor(greedy): make functional.
